Quantum Fiedler Eigenbasis for Rewiring Networks - QFERN Version 4
demonstrated here is the use of the Fiedler vector to perform a novel type of graph rewiring for optimization of Quantum DAG's. This method is dubbed QFERN, standing for Quantum Fiedler Eigenbasis for Rewiring Networks. It represents a method, implemented as a workflow that utilizes specifically the Fiedler vector and eigenbasis, to rewire or reconfigure quantum networks to optimize the bottlenecks in such networks, as represented by spectral gaps, and effective resistances in such graphs of the entanglement between nodes.